+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 #### Fixes
 
 * [#416](https://github.com/ruby-grape/grape-swagger/pull/416): Support recursive models - [@lest](https://github.com/lest).
+* [#418](https://github.com/ruby-grape/grape-swagger/pull/418): Replaced github ref to rubygems for external gems - [@Bugagazavr](https://github.com/Bugagazavr).
 
 ### 0.20.3 (May 9, 2016)